Restructurning app to server and client folders
