+++ server/controllers/task/delete.js
