+++ server/models/house/schema.js
