+++ server/tests/user.test.js
@@ -1,5 +1,5 @@
 const request = require('supertest');
-const app = require('../app');
+const app = require('./../../app');
 const User = require('../models/user');
 
 const {
